Remove some import * statements and clean up code.
This cleans up the code but does not help with the import * issue
mentioned in #578.

When the following import statements are removed, the tests start to
fail with "no heavy atom" errors:

* aa.py - `from .structures import *`
* protein.py - `from .aa import *`

Some routines (particularly in utilities.py) have also been flagged with
TODO comments as candidates for replacement with numpy.  See #565.